feature:ui: add afterload request
