+++ src/ui/src/setup/afterload.js
@@ -1,0 +1,24 @@
+export function getBusiness (app) {
+    return app.$store.dispatch('objectBiz/searchBusiness', {
+        params: {
+            'fields': ['bk_biz_id', 'bk_biz_name'],
+            'condition': {
+                'bk_data_status': {
+                    '$ne': 'disabled'
+                }
+            }
+        },
+        config: {
+            requestId: 'post_searchBusiness_$ne_disabled'
+        }
+    }).then(business => {
+        app.$store.commit('objectBiz/setBusiness', business.info)
+        return business
+    })
+}
+
+export default async function (app) {
+    return Promise.all([
+        getBusiness(app)
+    ])
+}
