+++ src/ui/src/setup/preload.js
@@ -34,26 +34,6 @@ export function getAuthorizedBusiness (app) {
     return app.$store.dispatch('objectBiz/getAuthorizedBusiness')
 }
 
-export function getBusiness (app) {
-    return app.$store.dispatch('objectBiz/searchBusiness', {
-        params: {
-            'fields': ['bk_biz_id', 'bk_biz_name'],
-            'condition': {
-                'bk_data_status': {
-                    '$ne': 'disabled'
-                }
-            }
-        },
-        config: {
-            ...preloadConfig,
-            requestId: 'post_searchBusiness_$ne_disabled'
-        }
-    }).then(business => {
-        app.$store.commit('objectBiz/setBusiness', business.info)
-        return business
-    })
-}
-
 export function getUserCustom (app) {
     return app.$store.dispatch('userCustom/searchUsercustom', {
         config: {
@@ -82,7 +62,6 @@ export default async function (app) {
     return Promise.all([
         getClassifications(app),
         getUserCustom(app),
-        getBusiness(app),
         getUserList(app)
     ])
 }
